rename scripts to be better
